server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
bcrypt==4.0.1
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
dataclasses-json==0.6.7
debugpy==1.8.18
decorator==5.2.1
distro==1.9.0
ecdsa==0.19.1
executing==2.2.1
fastapi==0.124.2
filetype==1.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.55.0
greenlet==3.3.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
langchain==1.1.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.3
langchain-google-genai==4.0.0
langchain-groq==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.2
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.15
langsmith==0.4.59
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
nest-asyncio==1.6.0
numpy==2.3.5
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
pgvector==0.4.2
pillow==12.0.0
platformdirs==4.5.1
postgres==4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
psycopg==3.3.2
psycopg-pool==3.3.0
psycopg2-binary==2.9.11
psycopg2-pool==1.2
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
streamlit==1.52.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.3
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
